+++ src/utils/mod.rs
@@ -1,3 +1,3 @@
 pub type Color = [f32; 4];
 
-pub type Position = (u16, u16);
+pub type Position = (u32, u32);
